PHP installer (#1335)
* draft of PHP installation script

* start validating options in php installer

* setup integration testing for the script

* rename installation script file

* installer: add functions to build the known commands names matrix

* installer: add function to search for relevant php binaries + tests

* installer: finalize function to read relevant INI values

* installer: add progress indication to download

* installer: add function to pick interactively the php binaries for installation

* installer: installation until extension copy

* installer: start adding ini files template

* installer: add comments to explain INI settings

* installer: remove indentation from heredoc in INI template

* installer: do not run if ddtrace installed + rename tracer-specific options

* installer: remove check to exit if ddtrace module is detected

* installer: move instead of copying the .so file to avoid a segfault

* installer: replace the link to the right dd_wrap_autoloader.php

* installer: use always the same temporarily extension name before renaming

* installer: fix INI replacement sed and indentation for phpcs

* installer: add the possibility to install to all binaries

* installer: add package verification to CI

* installer: fix linting

* installer: use small resources in CI for package verification and remove obsolete unused executor

* installer: do not use cache while installing libexecinfo

* installer: make installer compatible with php 7.2-

* installer: grab API version from PHP API ini info

* installer: fix proper installation/verification of new installer on debian

* installer: support apache2 conf.d in debian based envs

* installer: rename to dd-library-php-setup.php

* installer: add checks for libcurl and fallbacks while downloading the installable archive

* installer: support uninstall

* installer: test and fix upgrade process

* installer: test and fix upgrade process from php installer to php installer

* installer: test and fix uninstall process

* installer: test and fix library verification

* installer: test and fix libcurl library verification

* installer: add non interactive hint

* installer: fix typo in verified file name

* installer: add simple installation test

* installer: refactor suffix generation + add alpine

* installer: support custom installation directory

* installer: automatically detect installer tests

* installer: minor - redefine tests search pattern

* installer: minor - per version installation dir

* installer: run tests in CI

* installer: fix phpcs

* installer: run installer tests in CI

* installer: less strings, more constants

* installer: add functions doc blocks

* installer: fix naming and binary paths in installer verification

* installer: comments and minor DRY

* installer: simplify integrations settings list

* installer: minor comments

* installer: make installer compatible with php 5.4-5.5

* installer: test php installer on centos 6 in CI

* installer: assert installation from file works

* installer: assert versions in installer tests that missed it

* installer: minor refactorings and comments to dd-library-php-setup.php

* Update dd-library-php-setup.php


* Apply suggestions from code review


* installer: improvements to the installation script based on CR

* installer: test proper error message when user has no permissions

* installer: improve error handling

* installer: improve search of binaries

* installer: fix local file test now that we do not copy to a tar.gz with fixed naming

* installer: giveup on finding only executable binaries as the find -executable and -perm options are not portable

* installer: use exact .tar.gz archive name to install in debian/centos/alpine

* Apply suggestions from code review


* installer: fix fallback date format for installation directory

* installer: replace multiline // comments with /**/

* installer: from --php-bin=php to --php-bin php - for all params

* installer: test download fallbacks - no ext-curl and no curl cli

* installer: fix install paths on centos6 verification scripts

* installer: fix installation on debian with apach2 separate folder

* installer: remove leftover from debian execution no present in alpine images

* installer: use the new datadog.trace.request_init_hook param for package installers

* installer: avoid deprecation warning about ddtrace.request_init_hook when upgrading from legacy to new installer

* installer: Ensure ini scan directory exists

+++ dockerfiles/verify_packages/verify_rpm_centos6.sh
@@ -10,10 +10,19 @@ fi
 for phpVer in $(ls ${PHP_INSTALL_DIR}); do
     echo "Installing ddtrace on PHP ${phpVer}..."
     switch-php $phpVer
-    rpm -Uvh /build_src/build/packages/*.rpm
-    php --ri=ddtrace
 
-    # Uninstall the tracer
-    rpm -e datadog-php-tracer
-    rm -f /opt/datadog-php/etc/ddtrace.ini
+    # Installing dd-trace-php
+    INSTALL_TYPE="${INSTALL_TYPE:-php_installer}"
+    if [ "$INSTALL_TYPE" = "native_package" ]; then
+        echo "Installing dd-trace-php using the OS-specific package installer"
+        rpm -Uvh /build_src/build/packages/*.rpm
+        php --ri=ddtrace
+
+        # Uninstall the tracer
+        rpm -e datadog-php-tracer
+        rm -f /opt/datadog-php/etc/ddtrace.ini
+    else
+        echo "Installing dd-trace-php using the new PHP installer"
+        php /build_src/dd-library-php-setup.php --tracer-file /build_src/build/packages/*.tar.gz --php-bin all
+    fi
 done
